🎨 refactor: バーコード読み取り画面のリファクタ & スタイル調整
